requirements.txt
SQLAlchemy==2.0.36
python-dotenv==1.0.1
aiogram==3.15.0